moving test cases to specific files to avoid the too long encode_decode_api_test.cpp
