+++ test/api/targets.mk
@@ -4,11 +4,16 @@ API_TEST_CPP_SRCS=\
 	$(API_TEST_SRCDIR)/BaseEncoderTest.cpp\
 	$(API_TEST_SRCDIR)/cpp_interface_test.cpp\
 	$(API_TEST_SRCDIR)/DataGenerator.cpp\
+	$(API_TEST_SRCDIR)/decode_api_test.cpp\
 	$(API_TEST_SRCDIR)/decode_encode_test.cpp\
+	$(API_TEST_SRCDIR)/decoder_ec_test.cpp\
 	$(API_TEST_SRCDIR)/decoder_test.cpp\
 	$(API_TEST_SRCDIR)/encode_decode_api_test.cpp\
+	$(API_TEST_SRCDIR)/encode_options_test.cpp\
 	$(API_TEST_SRCDIR)/encoder_test.cpp\
+	$(API_TEST_SRCDIR)/ltr_test.cpp\
 	$(API_TEST_SRCDIR)/simple_test.cpp\
+
 
 API_TEST_OBJS += $(API_TEST_CPP_SRCS:.cpp=.$(OBJ))
